update universal version
Please pay attention: the codes might go wrong cause the pod file is
generated at a specific directory. When you download the codes, unzip
it and open in, there might be an error like that:
ld: library not found for -lAFNetworking
clang: error: linker command failed with exit code 1 (use -v to see
invocation)

In that case you need to regenerate the whole workspace and Pods
project.